backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
supabase==2.3.4
google-generativeai==0.3.2
pydantic==2.6.1